+++ CHANGELOG.md
@@ -6,14 +6,17 @@ Version counting is based on semantic versioning (Major.Feature.Patch)
 
 ### YACReader
 * Fixed color selection dialog appearing as a subwindow in macos.
+* Better support for HDPI screens (SVG icons).
 
 ### YACReaderLibrary
 * Update QtWebApp webserver to v1.8.6
 * New folder content view that replaces the old `subfolders in this folder` view shown when folders don't have direct comics.
-* Fixed going forward history navigation.
+* Fixed going forward in history navigation.
 * Continue Reading view that it is shown for the root folder.
 * UI gets updated when YACReaderLibrary gets updates from YACReader or YACReader for iOS.
 * Linux: Add fallback for dynamically loading libqrencode on distros that don't provide unversioned library symlinks
+* Fixed selected folder restoration after folder updates.
+* Better support for HDPI screens (SVG icons).
 
 ### Server
 * Add webui status page (reachable by navigating to server::port/webui)